Fixes for compiler warnings
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@7888 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/NikonTE2000/TEHub.cpp
@@ -915,7 +915,7 @@ int TEHub::ParseResponse(MM::Device& device, MM::Core& core, const char* cmd, st
       return DEVICE_SERIAL_INVALID_RESPONSE;
 
    value = rcvBuf_ + 4;
-   unsigned int responselength_ = strlen(rcvBuf_);
+   size_t responselength_ = strlen(rcvBuf_);
    if (rcvBuf_[0] == 'n' && responselength_ > 4)
    {
       int err = atoi(value.c_str()); // error occured
